Moved the Core/utility/Translators to the CQL Plugin. Modified CR plugin to leverage CQL plugin and it's configuration.
