+++ plugin/cpg/src/main/resources/application.yaml
@@ -5,5 +5,3 @@ hapi:
       log_enabled: false
       cql_debug_enabled: false
       use_embedded_translator_content: true
-      cqlTranslatorOptions:
-         analyzeDataRequirements: true
