+++ plugin/cql/src/main/resources/application.yaml
@@ -3,5 +3,7 @@ hapi:
     cql:
       enabled: true
       use_embedded_translator_content: true
+      cql_logging_enabled: false
       cqlTranslatorOptions:
         analyzeDataRequirements: true
+        collapseDataRequirements: true
